Sprint 753: ADR-018 + recon_engine pilot relocation
ADR-018 documents the per-tool domain package target layout
(services/audit/<tool>/{analysis,schemas,export}.py) and adopts
ADR-013's incremental migration discipline — one engine per sub-sprint,
backward-compat shims preserve existing imports, no big-bang.

Pilot: recon_engine.py (193 lines, 4 public symbols) moved to
services/audit/flux/recon.py. The legacy backend/recon_engine.py is
now a ~25-line shim re-exporting ReconEngine, ReconResult, ReconScore,
RiskBand. All 6 existing callers work without modification.

Why recon: small, self-contained (only depends on flux_engine.FluxResult
+ security_utils), pairs with the existing services/audit/flux_service.py
orchestrator that already lived in the audit services namespace.

Shim contract test (3 tests) asserts symbol identity between shim and
canonical paths — guards against future renames silently breaking the
~6 callers.

60/60 tests pass across test_recon_engine, test_recon_engine_shim,
test_audit_flux_routes, test_export_routes.

The remaining ~33 top-level *_engine.py files are deferred to
sub-sprints per ADR-018's incremental discipline. A lint scan that
flags them is folded into Sprint 756.
